emoroberta_backend/requirements.txt
fastapi==0.99.1
uvicorn==0.27.1
tensorflow==2.13.0
transformers==4.38.2
huggingface_hub
python-dotenv